Detect request timeout and display an error
This is now necessary because recently, the app has started randomly quietly failing to update some items. Since error logging had only been to the console, it went unseen. This correlates with changes to the library's WiFi, leading me to believe that the new WiFi is less reliable.
I also changed postitem.php to deliberately delay if the note is "zzz"; this is a way of testing the app.